Adding DefaultTaxEntity entity
