+++ src/entities/OrderTax.java
@@ -1,20 +1,9 @@
 package entities;
 
-public class OrderTax extends BaseEntity<Long>{
-    private Tax tax;
+public class OrderTax extends DefaultTaxHistory {
     private Order order;
-    private String taxName;
-    private Double taxPercentage;
 
     public OrderTax() {
-    }
-
-    public Tax getTax() {
-        return tax;
-    }
-
-    public void setTax(Tax tax) {
-        this.tax = tax;
     }
 
     public Order getOrder() {
@@ -23,21 +12,5 @@ public class OrderTax extends BaseEntity<Long>{
 
     public void setOrder(Order order) {
         this.order = order;
-    }
-
-    public String getTaxName() {
-        return taxName;
-    }
-
-    public void setTaxName(String taxName) {
-        this.taxName = taxName;
-    }
-
-    public Double getTaxPercentage() {
-        return taxPercentage;
-    }
-
-    public void setTaxPercentage(Double taxPercentage) {
-        this.taxPercentage = taxPercentage;
     }
 }
